Handling ACL in different actions of admin panel

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -10,6 +10,7 @@ use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\DB;
 use App\Permission;
 use Session;
+use Auth;
 
 class PermissionController extends Controller
 {
@@ -21,7 +22,8 @@ class PermissionController extends Controller
     public function index()
     {
         // $AllPermission = DB::table('permissions')->get();
-        $AllPermission=Permission::get();
+        $AllPermission = Permission::orderBy('id', 'asc')->paginate(5);
+        // $AllPermission=Permission::get();
         return view('admin/permission.permission')->withAllpermission($AllPermission);
     }
 
@@ -43,17 +45,24 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
-        $this->validate($request, array(
-                'name'         => 'required|max:255',
-                'label'        => 'required|max:1000',
-            ));
-        // store in the database
-        $permission = new Permission;
-        $permission->name = $request->name;
-        $permission->label = $request->label;
-        $permission->save();
-        Session::flash('success', 'The permission was successfully save!');
-        return redirect()->route('permissions.index');
+        if (Auth::guard('admin')->user()->can('create_permission')) { 
+            $this->validate($request, array(
+                    'name'         => 'required|max:255',
+                    'label'        => 'required|max:1000',
+                ));
+            // store in the database
+            $permission = new Permission;
+            $permission->name = $request->name;
+            $permission->label = $request->label;
+            $permission->save();
+            Session::flash('success', 'The permission was successfully save!');
+            return redirect()->route('permissions.index');
+        }
+        else {
+            Session::flash('error', 'You are not Authorised');
+            $AllPermission = Permission::orderBy('id', 'asc')->paginate(5);
+            return view('admin/permission.permission')->withAllpermission($AllPermission);
+        }
     }
 
     /**
@@ -75,9 +84,16 @@ class PermissionController extends Controller
      */
     public function edit($id)
     {
-        $AllPermission = DB::table('permissions')->get();
+       if (Auth::guard('admin')->user()->can('edit_permission')) { 
+        $AllPermission = Permission::orderBy('id', 'asc')->paginate(5);
         $permission = Permission::find($id);
         return view('admin/permission.permission')->withPermission($permission)->withAllpermission($AllPermission);
+        }
+        else {
+            Session::flash('error', 'You are not Authorised');
+            $AllPermission = Permission::orderBy('id', 'asc')->paginate(5);
+            return view('admin/permission.permission')->withAllpermission($AllPermission);
+        }
     }
 
     /**
@@ -111,8 +127,15 @@ class PermissionController extends Controller
      */
     public function destroy($id)
     {
-        $permission = Permission::findOrFail($id);
-        $permission->delete();
-        return redirect()->route('permissions.index');
+        if (Auth::guard('admin')->user()->can('edit_permission')) { 
+            $permission = Permission::findOrFail($id);
+            $permission->delete();
+            return redirect()->route('permissions.index');
+        }
+        else {
+            Session::flash('error', 'You are not Authorised');
+            $AllPermission = Permission::orderBy('id', 'asc')->paginate(5);
+            return view('admin/permission.permission')->withAllpermission($AllPermission);
+        }    
     }
 }
